refactor: fixed project files lineage endpoint packaging
