+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/projects/files/lineage/LineageSpec.scala
@@ -16,16 +16,16 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.lineage
+package io.renku.knowledgegraph.projects.files.lineage
 
+import LineageGenerators._
 import cats.syntax.all._
 import eu.timepit.refined.auto._
 import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators._
 import io.renku.graph.model.Schemas.schema
 import io.renku.graph.model.entities.{Activity, Entity}
-import io.renku.knowledgegraph.lineage.LineageGenerators._
-import io.renku.knowledgegraph.lineage.model.{Edge, Lineage, Node}
+import model.{Edge, Lineage, Node}
 import org.scalacheck.Gen
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
