+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/projects/files/lineage/NodeIdSpec.scala
@@ -16,17 +16,25 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.lineage
+package io.renku.knowledgegraph.projects.files.lineage
 
-import io.renku.knowledgegraph.docs.OpenApiTester._
+import LineageGenerators._
+import io.renku.generators.Generators.Implicits._
+import io.renku.graph.model.views.RdfResource
+import model.Node
+import org.apache.jena.util.URIref
+import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
+import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
-class EndpointDocsSpec extends AnyWordSpec {
+class NodeIdSpec extends AnyWordSpec with ScalaCheckPropertyChecks with should.Matchers {
 
-  "path" should {
+  "showAs[RdfResource]" should {
 
-    "return a valid Path object" in {
-      validatePath(EndpointDocs.path)
+    "wrap the Node.Id in <>" in {
+      forAll { resourceId: Node.Id =>
+        resourceId.showAs[RdfResource] shouldBe s"<${URIref.encode(resourceId.value)}>"
+      }
     }
   }
 }
